+++ Libplanet/Net/Messages/Message.cs
@@ -299,7 +299,17 @@ namespace Libplanet.Net.Messages
             message.Remote = remotePeer;
             message.Timestamp = timestamp;
 
-            if (!message.Remote.PublicKey.Verify(body.ToByteArray(), signature))
+            var headerWithoutSign = new[]
+            {
+                remains[(int)MessageFrame.Version],
+                remains[(int)MessageFrame.Type],
+                remains[(int)MessageFrame.Peer],
+                remains[(int)MessageFrame.Timestamp],
+            };
+
+            var messageForVerify = headerWithoutSign.Concat(body).ToArray();
+
+            if (!message.Remote.PublicKey.Verify(messageForVerify.ToByteArray(), signature))
             {
                 throw new InvalidMessageException("The message signature is invalid", message);
             }
@@ -348,11 +358,16 @@ namespace Libplanet.Net.Messages
             }
 
             // Write headers. (inverse order)
-            message.Push(key.Sign(message.ToByteArray()));
             message.Push(timestamp.Ticks);
             message.Push(SerializePeer(peer));
             message.Push((byte)Type);
             message.Push(version.Token);
+
+            // Make and insert signature
+            byte[] signature = key.Sign(message.ToByteArray());
+            List<NetMQFrame> frames = message.ToList();
+            frames.Insert((int)MessageFrame.Sign, new NetMQFrame(signature));
+            message = new NetMQMessage(frames);
 
             if (Identity is byte[] to)
             {